# controller/sheet.go
package controller

import (
	"errors"
	"time"

	"github.com/gin-gonic/gin"
	reqV1 "github.com/muxi-Infra/FeedBack-Backend/api/request/v1"
	"github.com/muxi-Infra/FeedBack-Backend/api/response"
	respV1 "github.com/muxi-Infra/FeedBack-Backend/api/response/v1"
	"github.com/muxi-Infra/FeedBack-Backend/domain"
	"github.com/muxi-Infra/FeedBack-Backend/errs"
	"github.com/muxi-Infra/FeedBack-Backend/pkg/ijwt"
	"github.com/muxi-Infra/FeedBack-Backend/service"
)

type SheetV1Handler interface {
	CreateTableRecord(c *gin.Context, r reqV1.CreatTableRecordReg, uc ijwt.UserClaims) (response.Response, error)
	GetTableRecordReqByKey(c *gin.Context, r reqV1.GetTableRecordReq, uc ijwt.UserClaims) (response.Response, error)
	GetTableRecordReqByRecordID(c *gin.Context, r reqV1.GetTableRecordByRecordIDReq, uc ijwt.UserClaims) (response.Response, error)
	GetFAQResolutionRecord(c *gin.Context, r reqV1.GetFAQProblemTableRecordReg, uc ijwt.UserClaims) (response.Response, error)
	UpdateFAQResolutionRecord(c *gin.Context, r reqV1.FAQResolutionUpdateReq, uc ijwt.UserClaims) (response.Response, error)
	GetPhotoUrl(c *gin.Context, r reqV1.GetPhotoUrlReq, uc ijwt.UserClaims) (res response.Response, err error)
}

type SheetV1 struct {
	s service.SheetService
	m service.MessageService
	a service.AuthService
}

func NewSheet(s service.SheetService, m service.MessageService, a service.AuthService) SheetV1Handler {
	sheet := &SheetV1{
		s: s,
		m: m,
		a: a,
	}

	return sheet
}

// CreateTableRecord 创建多维表格记录
//
//	@Summary		创建反馈记录
//	@Description	向指定的多维表格应用中添加用户反馈记录，支持文本内容、截图附件和联系方式。创建成功后会异步发送通知消息。
//	@Tags			Sheet
//	@ID				create-table-record
//	@Accept			json
//	@Produce		json
//	@Param			Authorization	header		string												true	"Bearer Token"
//	@Param			request			body		reqV1.CreatTableRecordReg							true	"新增记录请求参数"
//	@Success		200				{object}	response.Response{data=respV1.CreatTableRecordResp}	"成功返回创建记录结果"
//	@Failure		400				{object}	response.Response									"请求参数错误或飞书接口调用失败"
//	@Failure		500				{object}	response.Response									"服务器内部错误"
//	@Router			/api/v1/sheet/records [post]
func (s *SheetV1) CreateTableRecord(c *gin.Context, r reqV1.CreatTableRecordReg, uc ijwt.UserClaims) (response.Response, error) {
	if _, err := tableConfigWithScope(s.a, uc, "feedback:create"); err != nil {
		return response.Response{}, err
	}

	err := validateTableIdentify(*r.TableIdentify, uc.TableIdentity)
	if err != nil {
		return response.Response{}, err
	}

	// 组装参数
	record, err := buildCreateTableRecord(r, uc.StudentID)
	if err != nil {
		return response.Response{}, err
	}

	// 添加默认字段值，不依赖于飞书的默认值以及前端的传入
	t := time.Now()
	record.Record["进度"] = "待处理"
	record.Record["提交时间"] = t.UnixMilli()

	tableConfig, err := tableConfigFromClaims(s.a, uc)
	if err != nil {
		return response.Response{}, err
	}

	// 发起请求
	createdRecordID, err := s.s.CreateLarkRecord(record, &tableConfig)
	if err != nil {
		return response.Response{}, err
	}

	if createdRecordID == nil {
		return response.Response{
			Code:    0,
			Message: "Success",
			Data:    "",
		}, nil
	}

	// TODO 后续想改成 kafka 异步处理
	go func(recordID, content string, tc domain.TableConfig) {
		recordDate, url, err := s.s.GetTableRecordReqByRecordID(&recordID, &tc)
		if err != nil || url == nil {
			return
		}
		// 先保存数据库，再根据表格配置决定是否发送通知，避免通知失败导致记录丢失。
		err = s.s.CreateDBRecord(&recordID, url, recordDate, tc)
		if err != nil {
			return
		}
		if !tc.Notice {
			return
		}
		if err = s.m.SendLarkNotification(*tc.TableName, content, *url); err != nil {
			return
		}
	}(*createdRecordID, *r.Content, tableConfig)

	resp := respV1.CreatTableRecordResp{
		RecordID: *createdRecordID,
	}

	return response.Response{
		Code:    0,
		Message: "Success",
		Data:    resp,
	}, nil
}

// GetTableRecordReqByKey 获取用户历史反馈记录
//
//	@Summary		查询历史反馈记录
//	@Description	根据当前 JWT 中的学生身份查询历史反馈记录，支持分页查询。前端不再传入学生 ID。
//	@Tags			Sheet
//	@ID				get-table-record
//	@Accept			json
//	@Produce		json
//	@Param			Authorization	header		string												true	"Bearer Token"
//	@Param			request			query		reqV1.GetTableRecordReq								true	"查询记录请求参数"
//	@Success		200				{object}	response.Response{data=respV1.GetTableRecordResp}	"成功返回查询结果"
//	@Failure		400				{object}	response.Response									"请求参数错误或飞书接口调用失败"
//	@Failure		500				{object}	response.Response									"服务器内部错误"
//	@Router			/api/v1/sheet/records [get]
func (s *SheetV1) GetTableRecordReqByKey(c *gin.Context, r reqV1.GetTableRecordReq, uc ijwt.UserClaims) (response.Response, error) {
	if _, err := tableConfigWithScope(s.a, uc, "feedback:read:self"); err != nil {
		return response.Response{}, err
	}
	err := validateTableIdentify(*r.TableIdentify, uc.TableIdentity)
	if err != nil {
		return response.Response{}, err
	}

	// 组装参数
	keyField := domain.TableField{
		FieldName: r.KeyFieldName,
		Value:     &uc.StudentID,
	}
	tableConfig, err := tableConfigFromClaims(s.a, uc)
	if err != nil {
		return response.Response{}, err
	}

	serviceResult, err := s.s.GetTableRecordReqByKey(&keyField, r.RecordNames, r.PageToken, &tableConfig)
	if err != nil {
		return response.Response{}, err
	}

	resp := respV1.GetTableRecordResp{
		Records:   make([]domain.TableRecord, 0),
		HasMore:   false,
		PageToken: "",
	}

	if serviceResult.Records != nil {
		resp.Records = serviceResult.Records
	}
	if serviceResult.PageToken != nil {
		resp.PageToken = *serviceResult.PageToken
	}
	if serviceResult.HasMore != nil {
		resp.HasMore = *serviceResult.HasMore
	}

	return response.Response{
		Code:    0,
		Message: "Success",
		Data:    resp,
	}, nil
}

// GetTableRecordReqByRecordID 根据记录 ID 查询用户历史反馈记录
//
//	@Summary		按 RecordID 查询历史反馈记录
//	@Description	根据 record_id 获取单条用户反馈记录的详细内容，返回记录字段的键值对。请求中需包含合法的 table_identify，用于校验权限。
//	@Tags			Sheet
//	@ID				get-table-record-by-id
//	@Accept			json
//	@Produce		json
//	@Param			Authorization	header		string														true	"Bearer Token"
//	@Param			request			query		reqV1.GetTableRecordByRecordIDReq							true	"查询记录请求参数"
//	@Success		200				{object}	response.Response{data=respV1.GetTableRecordByRecordIdResp}	"成功返回查询结果"
//	@Failure		400				{object}	response.Response											"请求参数错误或飞书接口调用失败"
//	@Failure		500				{object}	response.Response											"服务器内部错误"
//	@Router			/api/v1/sheet/record [get]
func (s *SheetV1) GetTableRecordReqByRecordID(c *gin.Context, r reqV1.GetTableRecordByRecordIDReq, uc ijwt.UserClaims) (response.Response, error) {
	if _, err := tableConfigWithScope(s.a, uc, "feedback:read:self"); err != nil {
		return response.Response{}, err
	}
	err := validateTableIdentify(*r.TableIdentify, uc.TableIdentity)
	if err != nil {
		return response.Response{}, err
	}

	// 组装参数
	tableConfig, err := tableConfigFromClaims(s.a, uc)
	if err != nil {
		return response.Response{}, err
	}

	if err := s.s.VerifyTableRecordOwnership(r.RecordID, &uc.StudentID, &tableConfig); err != nil {
		return response.Response{}, err
	}

	serviceResult, _, err := s.s.GetTableRecordReqByRecordID(r.RecordID, &tableConfig)
	if err != nil {
		return response.Response{}, err
	}

	resp := respV1.GetTableRecordByRecordIdResp{
		Record: make(map[string]any),
	}

	if serviceResult != nil {
		resp.Record = serviceResult
	}

	return response.Response{
		Code:    0,
		Message: "Success",
		Data:    resp,
	}, nil
}

// GetFAQResolutionRecord 获取常见问题及解决状态
//
//	@Summary		查询FAQ问题记录
//	@Description	根据当前 JWT 中的学生身份查询相关的常见问题记录及其解决状态。
//	@Tags			Sheet
//	@ID				get-faq-resolution-record
//	@Accept			json
//	@Produce		json
//	@Param			Authorization	header		string														true	"Bearer Token"
//	@Param			request			query		reqV1.GetFAQProblemTableRecordReg							true	"查询记录请求参数"
//	@Success		200				{object}	response.Response{data=respV1.GetTableRecordByRecordIdResp}	"成功返回单条记录的字段键值对"
//	@Failure		400				{object}	response.Response											"请求参数错误或飞书接口调用失败"
//	@Failure		500				{object}	response.Response											"服务器内部错误"
//	@Router			/api/v1/sheet/records/faq [get]
func (s *SheetV1) GetFAQResolutionRecord(c *gin.Context, r reqV1.GetFAQProblemTableRecordReg, uc ijwt.UserClaims) (response.Response, error) {
	if _, err := tableConfigWithScope(s.a, uc, "feedback:read"); err != nil {
		return response.Response{}, err
	}
	err := validateTableIdentify(*r.TableIdentify, uc.TableIdentity)
	if err != nil {
		return response.Response{}, err
	}

	// 组装参数
	tableConfig, err := tableConfigFromClaims(s.a, uc)
	if err != nil {
		return response.Response{}, err
	}

	if err := validateStudentID(uc.StudentID); err != nil {
		return response.Response{}, err
	}
	faqServiceResult, err := s.s.GetFAQProblemTableRecord(&uc.StudentID, r.RecordNames, &tableConfig)
	if err != nil {
		return response.Response{}, err
	}

	resp := respV1.GetFAQProblemTableRecordResp{
		Records: make([]domain.FAQTableRecord, 0),
		Total:   0,
	}

	if faqServiceResult.Records != nil {
		resp.Records = faqServiceResult.Records
	}
	if faqServiceResult.Total != nil {
		resp.Total = *faqServiceResult.Total
	}

	return response.Response{
		Code:    0,
		Message: "Success",
		Data:    resp,
	}, nil
}

// UpdateFAQResolutionRecord 更新FAQ问题解决状态
//
//	@Summary		标记FAQ问题解决状态
//	@Description	用户更新FAQ问题的解决状态，将问题标记为已解决或未解决。
//	@Tags			Sheet
//	@ID				update-faq-resolution
//	@Accept			json
//	@Produce		json
//	@Param			Authorization	header		string							true	"Bearer Token"
//	@Param			request			body		reqV1.FAQResolutionUpdateReq	true	"更新FAQ解决状态请求参数"
//	@Success		200				{object}	response.Response				"成功更新FAQ解决状态"
//	@Failure		400				{object}	response.Response				"请求参数错误或飞书接口调用失败"
//	@Failure		500				{object}	response.Response				"服务器内部错误"
//	@Router			/api/v1/sheet/records/faq [post]
func (s *SheetV1) UpdateFAQResolutionRecord(c *gin.Context, r reqV1.FAQResolutionUpdateReq, uc ijwt.UserClaims) (response.Response, error) {
	if _, err := tableConfigWithScope(s.a, uc, "feedback:write"); err != nil {
		return response.Response{}, err
	}

	err := validateTableIdentify(*r.TableIdentify, uc.TableIdentity)
	if err != nil {
		return response.Response{}, err
	}

	// 组装参数
	FAQResolution := domain.FAQResolution{
		UserID:              &uc.StudentID,
		RecordID:            r.RecordID,
		ResolvedFieldName:   r.ResolvedFieldName,
		UnresolvedFieldName: r.UnresolvedFieldName,
		IsResolved:          r.IsResolved,
	}
	tableConfig, err := tableConfigFromClaims(s.a, uc)
	if err != nil {
		return response.Response{}, err
	}

	err = s.s.UpdateFAQResolutionRecord(&FAQResolution, &tableConfig)
	if err != nil {
		return response.Response{}, err
	}

	return response.Response{
		Code:    0,
		Message: "Success",
		Data:    nil,
	}, nil
}

// GetPhotoUrl 获取截图临时下载链接
//
//	@Summary		获取截图临时URL
//	@Description	根据文件Token列表批量获取截图的临时下载URL，用于前端展示或下载图片。
//	@Tags			Sheet
//	@ID				get-photo-url
//	@Accept			json
//	@Produce		json
//	@Param			Authorization	header		string					true	"Bearer Token"
//	@Param			request			query		reqV1.GetPhotoUrlReq	true	"获取截图URL请求参数"
//	@Success		200				{object}	response.Response		"成功返回临时URL信息"
//	@Failure		400				{object}	response.Response		"请求参数错误或飞书接口调用失败"
//	@Failure		500				{object}	response.Response		"服务器内部错误"
//	@Router			/api/v1/sheet/photos/url [get]
func (s *SheetV1) GetPhotoUrl(c *gin.Context, r reqV1.GetPhotoUrlReq, uc ijwt.UserClaims) (response.Response, error) {
	// todo 目前没有检测这张照片属于这个用户，属于目前这个登陆的项目，即图片 Token 没有归属校验
	if _, err := tableConfigWithScope(s.a, uc, "feedback:read:self"); err != nil {
		return response.Response{}, err
	}
	photoUrlResult, err := s.s.GetPhotoUrl(r.FileTokens)
	if err != nil {
		return response.Response{}, err
	}

	photoUrlResponse := respV1.GetPhotoUrlResp{
		Files: photoUrlResult,
	}

	return response.Response{
		Code:    0,
		Message: "Success",
		Data:    photoUrlResponse,
	}, nil
}

func validateTableIdentify(a, b string) error {
	if a != b {
		return errs.TableIdentifierInvalidError(errors.New("table identify is not invalid"))
	}
	return nil
}

// buildCreateTableRecord 组装以及校验创建记录的参数
func buildCreateTableRecord(r reqV1.CreatTableRecordReg, studentID string) (*domain.TableRecord, error) {
	// 拷贝 ExtraRecord，避免修改调用方原始 map
	totalRecord := make(map[string]any, len(r.ExtraRecord)+4)
	for k, v := range r.ExtraRecord {
		totalRecord[k] = v
	}

	// 必填字段校验
	if err := validateStudentID(studentID); err != nil {
		return nil, err
	}
	totalRecord["学号"] = studentID
	if r.Content == nil {
		return nil, errs.CreateRecordEmptyContentError(errors.New("content is required"))
	} else if len(*r.Content) == 0 {
		return nil, errs.CreateRecordEmptyContentError(errors.New("content is required"))
	}
	totalRecord["反馈内容"] = *r.Content

	// 可选字段处理
	// 将图片文件 token 数组转换为 Feishu API 期望的对象数组: [{"file_token": "your_file_token"}]
	fileObjs := make([]map[string]string, 0, len(r.Images))
	for _, t := range r.Images {
		fileObjs = append(fileObjs, map[string]string{"file_token": t})
	}
	totalRecord["截图"] = fileObjs
	if r.ContactInfo != nil {
		totalRecord["联系方式（QQ/邮箱）"] = *r.ContactInfo
	}

	record := &domain.TableRecord{
		Record: totalRecord,
	}
	return record, nil
}

// validateStudentID 校验来自已验证 JWT 的学生身份，不接受前端传入的学号。
func validateStudentID(studentID string) error {
	if studentID == "" {
		return errs.CreateRecordEmptyStudentIDError(errors.New("student_id is missing from token"))
	}
	if len(studentID) != 10 {
		return errs.CreateRecordInvalidStudentIDError(errors.New("student_id is invalid, length must be 10"))
	}
	return nil
}